store invalid descriptor when serial port is closed
The value of the fd property when the device is closed could be null, 0,
or undefined.  Most tests for open checked whether fd > 0, but one
checked fd >= 0 which would pass with all three closed values, and one
checked for strict equality with null which would pass with only one
closed value.

0 is a valid descriptor, though unlikely to be available in a node
procress.  -1 is outside the domain of valid descriptors so use it
instead.  Also use isOpen() instead of direct checks of the descriptor
when checking whether the instance is open.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>