serialport_unix: more cleanup
Remove the return value from setup(), as fact-of-error is indicated to
the caller by a non-NUL character at the start of the baton errorString.

Remove all close() calls from setup() as they were inconsistently
present.

Close the descriptor only when setup() failed on open, allowing update
failures to leave the serial port usable.

Correct the check for invoking close() with an invalid descriptor, but
continue to treat that as not-an-error.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>